rustc: streamline the Print/fmt::Display impls in ppaux and move them to ty::print::pretty.
